% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_size.R
\name{ci_size}
\alias{ci_size}
\title{Confidence Interval (Size)}
\usage{
ci_size(confidence_level, sample_size, mean, pop_sd, unknown_pop_sd = TRUE)
}
\arguments{
\item{confidence_level}{value, e.g. \code{0.95} or \code{95} for 95\% confidence level}

\item{sample_size}{integer, e.g. \code{30}}

\item{mean}{integer, e.g. \code{10}}

\item{pop_sd}{value, e.g. \code{3.14}}

\item{unknown_pop_sd}{boolean, e.g. \code{FALSE} (defaults to \code{TRUE}  otherwise)}
}
\description{
\code{ci_size} returns the confidence interval given a confidence level, sample size, mean and standard deviation.
}
\details{
By default, the computation assumes that sample standard deviation is used - if using population standard deviation (where it is known or given), set \code{unknown_pop_sd} to \code{FALSE}.

\code{confidence_level} accepts values for both frequency as well as probabilities.
}
\examples{
ci_size(0.95,30,10,3,TRUE)
}
