% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo.R
\name{hypo}
\alias{hypo}
\title{Hypothesis Testing}
\usage{
hypo(one_tail = TRUE, unknown_pop_sd = TRUE, x_bar, mu, sd, sample_size)
}
\arguments{
\item{one_tail}{boolean value for type of test, e.g. \code{FALSE} (defaults to \code{TRUE}  otherwise)}

\item{unknown_pop_sd}{boolean, e.g. \code{FALSE} (defaults to \code{TRUE}  otherwise)}

\item{x_bar}{value, e.g. \code{30}}

\item{mu}{value, e.g. \code{32}}

\item{sd}{value, e.g. \code{3.14}}

\item{sample_size}{integer, e.g. \code{30}}
}
\value{
Returns the \code{p-value}.
}
\description{
\code{hypo} returns test statistic (\emph{z or t-value}), location of \code{x_bar} on the graph (relative to centre), which test is used and the corresponding \code{p-value}.
}
\details{
By default, the computation assumes that sample standard deviation is used - if using population standard deviation (where it is known or given), set \code{unknown_pop_sd} to \code{FALSE}.

Similarly, the computation assumes that the one-tail test is used - if using the two-tail test, set \code{one-tail} to \code{FALSE}.
}
\examples{
hypo(TRUE, TRUE, 30, 32, 5, 30)
}
