% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{sd_pop}
\alias{sd_pop}
\title{Standard Deviation (Population)}
\usage{
sd_pop(x)
}
\arguments{
\item{x}{vector, e.g. \code{c(1,2,3,4)}}
}
\description{
\code{sd_pop} computes the standard deviation of the values in \code{x} with the denominator value \emph{n}, unlike the implementation as in \code{\link[stats]{sd}}.
}
\examples{
sd_pop(c(1,2,3,4))
}
\seealso{
\code{\link{var_pop}} for its square and  \code{\link[stats]{sd}} for the default \code{stats} package variant with denominator \emph{n-1}.
}
