% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{var_pop}
\alias{var_pop}
\title{Variance (Population)}
\usage{
var_pop(x)
}
\arguments{
\item{x}{vector, e.g. \code{c(1,2,3,4)}}
}
\description{
\code{var_pop} computes the variance of the values in \code{x} with the denominator value \emph{n}, unlike the implementation as in \code{\link[stats]{var}}.
}
\examples{
var_pop(c(1,2,3,4))
}
\seealso{
\code{\link{sd_pop}} for its root and  \code{\link[stats]{var}} for the default \code{stats} package variant with denominator \emph{n-1}.
}
