% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{weighted_stats}
\alias{weighted_stats}
\title{Key Statistics (Weighted)}
\usage{
weighted_stats(x, p)
}
\arguments{
\item{x}{vector of values, e.g. \code{c(2,3,4)}}

\item{p}{vector of probabilities/frequencies, e.g. \code{c(1,1,2)} or \code{c(0.25,0.25,0.5)}}
}
\value{
\itemize{
\item Expectation/Mean
\item Population Variance and Standard Deviation
}
}
\description{
\code{weighted_stats} returns the expectation/mean, population variance and standard deviations from a weighted vector (differing probabilities/distribution).
}
\details{
\code{probabilities} accepts values for both frequency as well as probabilities.
}
\examples{
weighted_stats(c(2,3,4), c(1,1,2))
weighted_stats(c(2,3,4), c(0.25,0.25,0.5))
}
\seealso{
\code{\link{key_stats}} for calculating x with a uniform distribution.
}
